Move discovery related classes to the discovery namespace.
